Move icons to icon folder
