+++ src/components/icons/Community.stories.tsx
@@ -1,8 +1,8 @@
 import { Meta, Story } from '@storybook/react';
 import React from 'react';
 
-import { CommunityIcon } from '../';
-import { SvgProps } from '../types';
+import { CommunityIcon } from '.';
+import { SvgProps } from './types';
 
 export default {
   component: CommunityIcon,
