+++ src/components/icons/DiscoveryIcon.stories.tsx
