+++ src/components/icons/DiscoveryIcon.tsx
