+++ src/components/icons/EmbeddedIcon.stories.tsx
