+++ src/components/icons/EmbeddedIcon.tsx
