+++ src/components/icons/EmpathyIcon.tsx
