+++ src/components/icons/ExecutionIcon.stories.tsx
