+++ src/components/icons/ExecutionIcon.tsx
