+++ src/components/icons/FullyManagedIcon.stories.tsx
