+++ src/components/icons/SustainabilityIcon.stories.tsx
