+++ src/components/icons/SustainabilityIcon.tsx
