+++ src/components/icons/index.ts
@@ -1,12 +1,12 @@
 export * from './LogoIcon';
 export * from './MenuIcon';
-export * from './custom-icons/CommunityIcon';
-export * from './custom-icons/DeliveryIcon';
-export * from './custom-icons/DiscoveryIcon';
-export * from './custom-icons/EmbeddedIcon';
-export * from './custom-icons/EmpathyIcon';
-export * from './custom-icons/ExecutionIcon';
-export * from './custom-icons/FullyManagedIcon';
-export * from './custom-icons/SupportIcon';
-export * from './custom-icons/SustainabilityIcon';
-export * from './custom-icons/TransparencyIcon';
+export * from './DeliveryIcon';
+export * from './CommunityIcon';
+export * from './DiscoveryIcon';
+export * from './EmbeddedIcon';
+export * from './EmpathyIcon';
+export * from './ExecutionIcon';
+export * from './FullyManagedIcon';
+export * from './SupportIcon';
+export * from './SustainabilityIcon';
+export * from './TransparencyIcon';
